Move test classes into separate folder
