+++ src/BikeSack/TestingClasses/HeadLightTest.java
@@ -1,3 +1,5 @@
+package BikeSack.TestingClasses;
+import BikeSack.BikeSack;
 
 public class HeadLightTest {
 	public static void main(String[] args) {
